JavaDoc changes

+++ src/main/java/com/sandip/vm/exceptions/NotFullPaidException.java
@@ -4,9 +4,9 @@ public class NotFullPaidException extends RuntimeException {
 
 	private static final long serialVersionUID = 8525755416792112144L;
 	private final String message;
-	private final long remaining;
+	private final double remaining;
 
-	public NotFullPaidException(String message, long remaining) {
+	public NotFullPaidException(String message, double remaining) {
 		this.message = message;
 		this.remaining = remaining;
 	}
@@ -16,7 +16,7 @@ public class NotFullPaidException extends RuntimeException {
 		return message + remaining;
 	}
 
-	public long getRemaining() {
+	public double getRemaining() {
 		return remaining;
 	}
